+++ pages/groups/group-public.php
@@ -266,7 +266,7 @@ function sel($index, $value)
                 }
                 data.data.forEach(function(d) {
                     suggest.append(
-                        `<a onclick="selectActivity(this)" data-id="${d.id.toString()}">${d.details.icon} ${d.details.plain}</a>`
+                        `<a  data-id="${d.id.toString()}">${d.details.icon} ${d.details.plain}</a>`
                     )
                 })
                 suggest.find('a')