+++ routes/activities.php
@@ -253,6 +253,7 @@ Route::get('/activities/view/([a-zA-Z0-9]*)', function ($id) {
     include_once BASEPATH . "/php/Render.php";
 
     $user = $_SESSION['username'];
+    $str_id = $id;
     $id = $DB->to_ObjectID($id);
     $activity = $osiris->activities->findOne(['_id' => $id], ['projection' => ['file' => 0]]);
     if (empty($activity)) {
@@ -330,6 +331,11 @@ Route::get('/activities/view/([a-zA-Z0-9]*)', function ($id) {
         ['$or' => [['source_id' => $id], ['target_id' => $id]]]
     )->toArray();
 
+    $connected_news = [];
+    if ($Settings->featureEnabled('news', true)) {
+        $connected_news = $osiris->news->find(['activities' => $str_id])->toArray();
+    }
+
     $guests_involved = false;
     $guests = [];
     if ($Settings->featureEnabled('guests')) {
@@ -353,7 +359,7 @@ Route::get('/activities/view/([a-zA-Z0-9]*)', function ($id) {
         $tagLabel = $Settings->tagLabel();
     }
 
-    $files = $osiris->uploads->find(['type' => 'activities', 'id' => strval($id)])->toArray();
+    $files = $osiris->uploads->find(['type' => 'activities', 'id' => $str_id])->toArray();
 
     // check user preference for activity view
     $activity_view = $_GET['view'] ?? $USER['activity_view'] ?? 'none';